split mesh generator into own file
