+++ CMakeLists.txt
@@ -152,6 +152,7 @@ set (SceneFiles
     "Source/Engine/Scene/Light.h"
     "Source/Engine/Scene/Camera.h"
     "Source/Engine/Scene/Mesh.h"
+    "Source/Engine/Scene/MeshGenerator.h"
     "Source/Engine/Scene/Material.h"
     "Source/Engine/Scene/Model.h"
     "Source/Engine/Scene/GameObject.h"
